phptags
sangrados

+++ app/View/Common/view.ctp
@@ -11,6 +11,8 @@ $line_add = $this->fetch('line_add');
 $line2_add = $this->fetch('line2_add');
 $from_controller = $this->fetch('from_controller');
 $from_id = $this->fetch('from_id');
+$button_edit_delete= $this->fetch ('button_edit_delete');
+$print_pdf= $this->fetch ('print_pdf');
 ?>
 
 <h2><?php echo $object; ?></h2>
@@ -22,29 +24,30 @@ $from_id = $this->fetch('from_id');
 </div>
 <div class="acciones">
 	<div class="printdet">
+	<?php if ($print_pdf !=0 or $print_pdf == null){?>
 	<ul><li>
-<?php 
-echo $this->element('imprimirV');
-?>	
-
+	<?php echo $this->element('imprimirV');?>
 	</li>
 	<li>
 <?php
-//si esta definido el $from_controller, ponemos un botón
-//que devuelve al controller mencionado,
-//si no es el caso, volvemos al index del controller del
-//borrado
-echo $this->Button->edit($controller,$id)
-	.' '.
+}
+if($button_edit_delete != 0 or $button_edit_delete == null){
 	//si esta definido el $from_controller, ponemos un botón
 	//que devuelve al controller mencionado,
 	//si no es el caso, volvemos al index del controller del
 	//borrado
-	(empty($from_controller) ? 
-	$this->Button->delete($controller,$id,$object)
-	//:$this->Button->deleteLine($controller,$id,$from_controller,$from_id,$object)
-	:$this->Button->deleteLine($controller,$id,$object)
-);
+	echo $this->Button->edit($controller,$id)
+	    .' '.
+	    //si esta definido el $from_controller, ponemos un botón
+	    //que devuelve al controller mencionado,
+	    //si no es el caso, volvemos al index del controller del
+	    //borrado
+	    (empty($from_controller) ?
+	    $this->Button->delete($controller,$id,$object)
+	    //:$this->Button->deleteLine($controller,$id,$from_controller,$from_id,$object)
+	    :$this->Button->deleteLine($controller,$id,$object)
+	);
+}
 ?>
 	</li>
 	</ul>
